requirements.txt
django
djangorestframework
pytest
numpy
reportlab
